+++ quill/test/UtilityTest.cpp
@@ -1,6 +1,7 @@
 #include "doctest/doctest.h"
 
 #include "quill/Utility.h"
+#include <cstdint>
 #include <cstring>
 #include <cstdint>